Rename helpers to utils; remove unused imports
