+++ cognigraph/gui/brain_visual.py
@@ -22,8 +22,8 @@ from vispy.visuals import Visual
 import vispy.visuals.transforms as vist
 from vispy.scene.visuals import create_visual_node
 
-from ..helpers.vispy_utils import (Colormap, color2vb, convert_meshdata,
-                                   wrap_properties, normalize)
+from ..utils.vispy_utils import (Colormap, color2vb, convert_meshdata,
+                                 wrap_properties, normalize)
 
 
 logger = logging.getLogger('cognigraph')
